Implemented the Animation of Explosions with larger radius
Created a class ExplosionSegment which is used to play the animation of the explosion segment according to its position like Top ,End, left end right end etc.
Created a method createExplosionSegments in gameMap class which creates the segments such that the animation is blocked by the indestructible wall,  This method is then triggered in the map's tick method when the bomb is to be exploded.